# spsyncapi/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  handlers.errorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
...
  handlers.restoreRunGetResponse:
    properties:
      file_transfers:
        items:
          $ref: '#/definitions/restorerun.FileTransferDetails'
        type: array
      pagination:
        $ref: '#/definitions/handlers.Pagination'
      restore_run:
        $ref: '#/definitions/restorerun.RunDetails'
    type: object
  handlers.restoreRunListResponse:
    properties:
      pagination:
        $ref: '#/definitions/handlers.Pagination'
      restore_runs:
        items:
          $ref: '#/definitions/restorerun.RunDetails'
        type: array
    type: object
  handlers.restoreRunStartResponse:
    properties:
      restore_run:
        $ref: '#/definitions/restorerun.RunDetails'
    type: object
...
  restorerun.FileTransferDetails:
    properties:
      end_at:
        type: string
      file_path:
        type: string
      start_at:
        type: string
    type: object
  restorerun.RunDetails:
    properties:
      end_at:
        type: string
      id:
        type: string
      job_id:
        type: string
      start_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: HTTP API server for the SPSync platform.
  title: SPSync API
  version: "1.0"
paths:
...
  /restore-jobs/{id}/runs:
    post:
      description: Creates a restore run row and starts the Temporal workflow
      parameters:
      - description: Restore job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.restoreRunStartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Start restore run
      tags:
      - restore-runs
  /restore-runs:
    get:
      description: Returns paginated restore runs; optional job_id filter
      parameters:
      - description: Filter by restore job ID
        in: query
        name: job_id
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.restoreRunListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List restore runs
      tags:
      - restore-runs
  /restore-runs/{id}:
    get:
      description: Returns a restore run and paginated file transfers
      parameters:
      - description: Restore run ID
        in: path
        name: id
        required: true
        type: string
      - description: File transfers page (default 1)
        in: query
        name: page
        type: integer
      - description: File transfers page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.restoreRunGetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get restore run
      tags:
      - restore-runs
  /restore-runs/{id}/stop:
    post:
      description: Cancels the Temporal workflow for an in-progress restore run
      parameters:
      - description: Restore run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.restoreRunStartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Stop restore run
      tags:
      - restore-runs
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Example: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"